Dropdowns complete. Note that dropdowns can easily be removed if they are deemed unecessary for the functionality of this webpage
